+++ db/migrate/20130625162942_rename_source_join_to_join.rb
@@ -1,0 +1,5 @@
+class RenameSourceJoinToJoin < ActiveRecord::Migration
+  def change
+    rename_table :source_joins, :joins
+  end
+end