+++ test/helpers/source_joins_helper_test.rb
@@ -1,4 +1,0 @@
-require 'test_helper'
-
-class SourceJoinsHelperTest < ActionView::TestCase
-end