+++ test/models/join_test.rb
@@ -1,0 +1,5 @@
+require 'test_helper'
+
+class JoinTest < ActiveSupport::TestCase
+
+end